Move JS (and static resources) to new folder
Spring Boot *really* wants to usr src/main/webapp to find static
resources, which does not work when transpiling JS and CSS files.
