+++ src/main/es/components/ActivityDialog.js
