+++ src/main/es/components/C3Graph.js
