+++ src/main/es/components/ChartPanel.js
