+++ src/main/es/components/DeleteDialog.js
