+++ src/main/es/utils/DistanceUtils.js
