+++ src/test/es/components/ErrorDialog.test.js
@@ -2,7 +2,7 @@
 
 import React from "react";
 import {shallow} from "enzyme";
-import ErrorDialog from "../../../main/webapp/components/ErrorDialog";
+import ErrorDialog from "../../../main/es/components/ErrorDialog";
 
 describe("ErrorDialog", () => {
